requirements.txt
appnope==0.1.0
asgiref==3.2.10
backcall==0.1.0
bcrypt==3.1.7
cffi==1.13.2
decorator==4.4.1
Django==3.0.8
django-crispy-forms==1.8.0
djangorestframework==3.11.0
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pytz==2020.1
six==1.13.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.1.7
